+++ pkg/deployments/README.md
@@ -68,9 +68,9 @@ Returns an object with all contracts from a deployment and their addresses.
 | Batch Relayer                                        | [`20211203-batch-relayer`](./tasks/20211203-batch-relayer)                                           |
 | Linear Pools for Aave aTokens                        | [`20211208-aave-linear-pool`](./tasks/20211208-aave-linear-pool)                                     |
 | Batch Relayer V2                                     | [`20220318-batch-relayer-v2`](./tasks/20220318-batch-relayer-v2)                                     |
-| Authorizer Adaptor for extending governance          | [`20220325-authorizer-adaptor `](./tasks/20220325-authorizer-adaptor)                                |
-| Wallet for the BAL token                             | [`20220325-bal-token-holder-factory `](./tasks/20220325-bal-token-holder-factory)                    |
-| Admin of the BAL token                               | [`20220325-balancer-token-admin `](./tasks/20220325-balancer-token-admin)                            |
+| Authorizer Adaptor for extending governance          | [`20220325-authorizer-adaptor`](./tasks/20220325-authorizer-adaptor)                                 |
+| Wallet for the BAL token                             | [`20220325-bal-token-holder-factory`](./tasks/20220325-bal-token-holder-factory)                     |
+| Admin of the BAL token                               | [`20220325-balancer-token-admin`](./tasks/20220325-balancer-token-admin)                             |
 | Gauge Registrant                                     | [`20220325-gauge-adder`](./tasks/20220325-gauge-adder)                                               |
 | Liquidity Mining: veBAL, Gauge Controller and Minter | [`20220325-gauge-controller`](./tasks/20220325-gauge-controller)                                     |
 | Mainnet Staking Gauges                               | [`20220325-mainnet-gauge-factory`](./tasks/20220325-mainnet-gauge-factory)                           |
@@ -83,13 +83,14 @@ Returns an object with all contracts from a deployment and their addresses.
 | veBAL Smart Wallet Checker                           | [`20220420-smart-wallet-checker`](./tasks/20220420-smart-wallet-checker)                             |
 | Fee Distributor for veBAL holders                    | [`20220420-fee-distributor`](./tasks/20220420-fee-distributor)                                       |
 | Linear Pools for Unbutton tokens                     | [`20220425-unbutton-aave-linear-pool`](./tasks/20220425-unbutton-aave-linear-pool)                   |
-| Distribution Scheduler for reward tokens on gauges   | [`20220422-distribution-scheduler`](./tasks/20220422-distribution-scheduler)                         |
 | Relayer with the fix for the Double Entrypoint issue | [`20220513-double-entrypoint-fix-relayer`](./tasks/20220513-double-entrypoint-fix-relayer)           |
 | Protocol Fee Withdrawer                              | [`20220517-protocol-fee-withdrawer`](./tasks/20220517-protocol-fee-withdrawer)                       |
 | Child Chain Gauge Token Adder                        | [`20220527-child-chain-gauge-token-adder`](./tasks/20220527-child-chain-gauge-token-adder)           |
 | Preseeded Voting Escrow Delegation                   | [`20220530-preseeded-voting-escrow-delegation`](./tasks/20220530-preseeded-voting-escrow-delegation) |
 | Stable Pools V2 of up to 5 tokens                    | [`20220609-stable-pool-v2`](./tasks/20220609-stable-pool-v2)                                         |
 | Gauge Registrant V2, supporting new networks         | [`20220628-gauge-adder-v2`](./tasks/20220628-gauge-adder-v2)                                         |
+| Distribution Scheduler for reward tokens on gauges   | [`20220707-distribution-scheduler`](./tasks/20220707-distribution-scheduler)                         |
+
 
 ## Scripts
