+++ pkg/deployments/deployment-txs/goerli.json
@@ -45,5 +45,6 @@
   "0x55B18f514D7f0C6C886d0C23Ac72dF5D3B5F8850": "0xb6b77224dbcc967aa835559cae1e72326be9aa6a95fdef01d9b8b019d96c2f61",
   "0xD360B8afb3d7463bE823bE1Ec3c33aA173EbE86e": "0x71bdf2cb1d2cf4c1521d82f6821aa0bee2c144252c3ae0dd7d651cb5bbcbc860",
   "0x42bAF6db21250fa76d015621D2F6DF172858A5cb": "0x00a32364bd6b3d17765be8f260ae3817dece638fa65eacb3a0b4ce91c67d5c38",
-  "0x7AF980bDBc36D21CE228EfABCCA35707566A2be5": "0xf4d174c73337af3e4a8ccc2b8bed061da453aae54981339f6cc9d0ac48d39192"
+  "0x7AF980bDBc36D21CE228EfABCCA35707566A2be5": "0xf4d174c73337af3e4a8ccc2b8bed061da453aae54981339f6cc9d0ac48d39192",
+  "0x05a0BF0540F346b6Dac25550738343BEb51C0f65": "0x7c2d19d0869a29e27f14fd8d48ba252cb1ec1a0869aa9d7d1e573d135c6c9f2b"
 }
