+++ pkg/deployments/deployment-txs/mainnet.json
@@ -49,5 +49,6 @@
   "0xE061bF85648e9FA7b59394668CfEef980aEc4c66": "0x42c1def229cfdc5c8dd0ea312f35e5501375c1748b8afb16805de18710184608",
   "0xd45369c11870e2057D5be17Cc106d32Ea416F7c4": "0xdcd6eecbc2b1fe9b38554e945169875f044d6e172bd11706de7dadbfb9018bbd",
   "0x2fFB7B215Ae7F088eC2530C7aa8E1B24E398f26a": "0xbe7835791b0989f62b246c8fb8abd0c2ad42aefd21f2f7cff38d440a9a8dd9dc",
-  "0x9588c26142e345f1A0d005CfC0C6DF29A8Fa010C": "0xa13a909fad8dafac3f8d309418aa8dc0f036e47b3f52a9ba7fb47d905fac11fa"
+  "0x9588c26142e345f1A0d005CfC0C6DF29A8Fa010C": "0xa13a909fad8dafac3f8d309418aa8dc0f036e47b3f52a9ba7fb47d905fac11fa",
+  "0xBd35248F8325DD1cB2bBf9D01E80A6bb99a792Dd": "0xbbc57a524fb34090a46c20371d7b4aedab217a9928aabbcbec05621acb16bb0b"
 }
