--- pkg/deployments/tasks/20220422-distribution-scheduler/index.ts
@@ -1,6 +1,0 @@
-import Task from '../../src/task';
-import { TaskRunOptions } from '../../src/types';
-
-export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
-  await task.deployAndVerify('DistributionScheduler', [], from, force);
-};
