--- pkg/deployments/tasks/20220422-distribution-scheduler/output/kovan.json
@@ -1,4 +1,0 @@
-{
-  "DistributionScheduler": "0x73e8ceFB3FaD5D263f0Ac5b29539f782ef336184",
-  "timestamp": 1650680352524
-}
