--- pkg/deployments/tasks/20220422-distribution-scheduler/output/mainnet.json
@@ -1,4 +1,0 @@
-{
-  "DistributionScheduler": "0xbfD9769b061E57e478690299011A028194D66e3C",
-  "timestamp": 1650680247439
-}
