+++ pkg/deployments/tasks/20220707-distribution-scheduler/output/goerli.json
@@ -1,0 +1,4 @@
+{
+  "DistributionScheduler": "0x05a0BF0540F346b6Dac25550738343BEb51C0f65",
+  "timestamp": 1657206679856
+}
