+++ pkg/deployments/tasks/20220707-distribution-scheduler/output/mainnet.json
@@ -1,0 +1,4 @@
+{
+  "DistributionScheduler": "0xBd35248F8325DD1cB2bBf9D01E80A6bb99a792Dd",
+  "timestamp": 1657208395655
+}
